+++ src/main/java/com/netcracker/dao/impl/mapper/CategoryExpenseFamilyReportMapper.java
@@ -14,7 +14,7 @@ public class CategoryExpenseFamilyReportMapper implements RowMapper<CategoryExpe
 
         AbstractCategoryReport categoryExpenseReport =
                 new CategoryExpenseReport.Builder()
-                        .amount(resultSet.getLong("amount"))
+                        .amount(resultSet.getDouble("amount"))
                         .categoryExpense(CategoryExpense.getNameByKey(resultSet.getBigDecimal("category").toBigInteger()))
                         .userReference(resultSet.getBigDecimal("user_id").toBigInteger())
                         .build();
