+++ src/main/java/com/netcracker/dao/impl/mapper/CategoryExpensePersonalReportMapper.java
@@ -14,7 +14,7 @@ public class CategoryExpensePersonalReportMapper implements RowMapper<CategoryEx
 
         AbstractCategoryReport categoryExpenseReport =
                 new CategoryExpenseReport.Builder()
-                        .amount(resultSet.getLong("amount"))
+                        .amount(resultSet.getDouble("amount"))
                         .categoryExpense(CategoryExpense.getNameByKey(resultSet.getBigDecimal("category").toBigInteger()))
                         .build();
         return (CategoryExpenseReport) categoryExpenseReport;
